requirements.txt
# Requirements for The Refactoring Swarm
# Python 3.10 or 3.11 required

# LLM API
google-generativeai>=0.3.0

# Code Quality Tools
pylint>=3.0.0
pytest>=7.4.0
autopep8>=2.0.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# Optional but useful
black>=23.0.0
